\name{Sicily2008M}
\docType{data}
\alias{Sicily2008M}
\title{Mortality data for the 2008 male population of Sicily Region (Italy)}
\description{
 This data set contains the 2008 male population of Sicily Region (Italy) observed mortality rates \code{Sicily2008M$qx} and the exposed to the risk of death \code{Sicily2008M$ex}, at ages 0,1,...,100.}
\usage{Sicily2008M}
\format{A data.frame containing the numeric vectors \code{qx} and \code{ex} of lenght 101. Row names are the ages  0,1,...,100}
\source{Italian National Institute of Statistics, data downloable from http://demo.istat.it/.}
\references{
Mazza A, Punzo A (2013) Graduation by Adaptive Discrete Beta Kernels. In A Giusti, G Ritter, M Vichi (eds.), \emph{Classification and Data Mining}, volume \bold{44} of \emph{Studies in Classification, Data Analysis and Knowledge Organization}, pp. 77-84. Springer-Verlag, Berlin-Heidelberg.
}
\seealso{
\code{\link{DBKGrad-package}}, \code{\link{dbkGrad}}, \code{\link{Sicily2008M}}
}
\keyword{datasets}