\name{DBKGrad-package}
\alias{DBKGrad-package}
\docType{package}
\title{DBKGrad - Discrete Beta Kernel Graduation}
\description{
This package allows for nonparametric graduation of mortality rates using the discrete beta kernel estimator in both its fixed (Mazza and Punzo, 2011) and adaptive (Mazza and Punzo, 2013a, 2013b) variants.
}
\details{
\tabular{ll}{
Package: \tab DBKGrad\cr
Type: \tab Package\cr
Version: \tab 1.5\cr
Date: \tab 2014-04-08\cr
License: \tab GNU-2\cr
}
}
\references{
Elzhov TV, Mullen KM, Bolker B (2010) \emph{minpack.lm:} R Interface to the Levenberg-Marquardt Nonlinear Least-Squares Algorithm Found in MINPACK. R package version 1.1-5. URL http://CRAN.R-project.org/package=minpack.lm.

Bagnato L, Punzo A (2013) Finite Mixtures of Unimodal Beta and Gamma Densities and the k-Bumps Algorithm. \emph{Computational Statistics}, \bold{28} (4), pp. 1571-159. doi:710.1007/s00180-012-0367-4.

Mazza A, Punzo A (2011) Discrete Beta Kernel Graduation of Age-Specific Demographic Indicators. In S Ingrassia, R Rocci, M Vichi (eds.), \emph{New Perspectives in Statistical Modeling and Data Analysis}, \emph{Studies in Classification, Data Analysis and Knowledge Organization}, pp. 127-134. Springer-Verlag, Berlin-Heidelberg.

Mazza A, Punzo A (2013a) Graduation by Adaptive Discrete Beta Kernels. In A Giusti, G Ritter, M Vichi (eds.), \emph{Classification and Data Mining}, \emph{Studies in Classification, Data Analysis and Knowledge Organization}, pp. 77-84. Springer-Verlag, Berlin-Heidelberg.

Mazza A, Punzo A (2013b) Using the Variation Coefficient for Adaptive Discrete Beta Kernel Graduation. In P Giudici, S Ingrassia, M Vichi (eds.), \emph{Advances in Statistical Modelling for Data Analysis, Studies in Classification, Data Analysis and Knowledge Organization}, pp. 225-232, Springer International Publishing, Switzerland.

Mazza A, Punzo A (2014) DBKGrad: An R Package for Mortality Rates Graduation by Discrete Beta Kernel Techniques. \emph{Journal of Statistical Software, Code Snippets}, \bold{57}{2}, 1-18.

More J (1978) The Levenberg-Marquardt Algorithm: Implementation and Theory. In G Watson (ed.), \emph{Numerical Analysis}, volume \bold{630} of \emph{Lecture Notes in Mathematics}, pp. 104-116. Springer- Verlag, Berlin-Heidelberg.

Punzo A (2010) Discrete Beta-type Models. In H Locarek-Junge, C Weihs (eds.), \emph{Classification as a Tool for Research}, \emph{Studies in Classification, Data Analysis and Knowledge Organization}, pp. 253-261. Springer-Verlag, Berlin-Heidelberg.

Punzo A, Zini A (2012) Discrete Approximations of Continuous and Mixed Measures on a Compact Interval \emph{Statistical Papers}, \bold{53}(3), 563-575.
}
\author{
Angelo Mazza and Antonio Punzo

Maintainer: Angelo Mazza <a.mazza@unict.it>
}
%%\note{
%%  ~~further notes~~
%%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{dbkGrad}}, \code{\link[DBKGrad:plot.dbkGrad]{plot}}, \code{\link{ItalyM}}
}
